Add files via upload
Here is the major update, as promised! I wrote a Jupyter Notebook that reads the polling data for the 2020 Presidential election published by 538 on Github, sorts through it, and plots the polling results and trends at the national and state levels. This poll reader also allows you to account for bias in the polls and use the resulting trend to create a very naive election forecast.