Add files via upload
By 5:00 pm on 11/04/2020, it seems like the local state-level polls were biased in favor of Biden’s popular vote in the states by an average of 2.49%. At the same time, the state polls seem to be biased against Trump by an average of 2.29%  Could it be that the pollsters were weighing their samples incorrectly in favor of Biden or is it possible that the surge in participation threw off the likely voter models pollsters rely on? Here is the Jupyter-Notebook I wrote that reads the polling data for the 2020 Presidential election, sorts through it and plots the polling results – as well as the election outcome based on those state level polls. The final vote count is not in, and we won’t be sure until the last count is complete.