Add files via upload
Is it possible that the surge in voter participation threw off the polls' likely voter models? Here is a rerun based on real election returns vs. the 'likely-voter' polls. It seems the state-level polls were biased in favor of Biden by an average of 2.49%. At the same time, the state polls seem to be biased against Trump by an average of 2.29%   Here is the Jupyter-Notebook I wrote that reads the polling data for the 2020 Presidential election and plots the election outcome after adjusting for the polls' state-level bias. Interestingly, it has Trump winning Arizona.